move Partition method to mysql driver
