+++ go.mod
@@ -7,7 +7,6 @@ require (
 	github.com/clbanning/mxj/v2 v2.7.0
 	github.com/fatih/color v1.15.0
 	github.com/fsnotify/fsnotify v1.6.0
-	github.com/gogf/gf/contrib/drivers/mysql/v2 v2.5.4
 	github.com/gorilla/websocket v1.5.0
 	github.com/grokify/html-strip-tags-go v0.0.1
 	github.com/magiconair/properties v1.8.6
